pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "tomas cufaro",email = "t.cufarofernandez@xero.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "openai (>=1.68.2,<1.76.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "crewai (>=0.121.1,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
